chore: Move dataStructureUtil functions to its own file
Related Issue: #5
