--- src/utils/mapUtils.ts
@@ -1,11 +1,0 @@
-function createMapFromObject<T>(obj: T) {
-  const map = new Map();
-
-  for (let key in obj) {
-    map.set(key, obj[key]);
-  }
-
-  return map;
-}
-
-export { createMapFromObject };
